+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Remove fields from database for design entity. [#549](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/549)
+- Remove emails from researcher entity. [#593](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/593)
 
 ## [2.17.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/37) 2023-11-02